remove UnloadedGlyphFactory
git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@11262 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ plugins/GraphAdjuster/src/com/affymetrix/igb/graph/SimpleGraphTab.java
@@ -969,11 +969,9 @@ public final class SimpleGraphTab
 
 			GlyphI parentgl = gl.getParent();
 			parentgl.removeChild(gl);
-			if (parentgl.getChildCount() == 0) {  // if no children left in tier, then remove it
-				if (isTierGlyph(parentgl)) {
-					igbService.deleteGlyph(parentgl);
-					igbService.packMap(false, false);
-				}
+			if (isTierGlyph(parentgl) && ((TierGlyph)parentgl).isGarbage()) {  // if no children left in tier, then remove it
+				igbService.deleteGlyph(parentgl);
+				igbService.packMap(false, false);
 			}
 		} else {
 			igbService.deleteGraph(gsym);
